manage change lists
