+++ src/components/manage/LinkList.tsx
@@ -7,11 +7,11 @@ const LinkListItem = () => {}
 export const LinkList = ({
   items,
   active,
-  setActive,
+  onClick,
 }: {
   items: LinkType[]
   active: LinkType | null
-  setActive: Dispatch<SetStateAction<LinkType | null>>
+  onClick: (link: LinkType) => void
 }) => {
   return (
     <div className=" border-b-2 lg:basis-1/5 lg:border-b-0 lg:border-r-2">
@@ -21,7 +21,7 @@ export const LinkList = ({
             active?.slug == folder.slug ? "bg-gray-100" : "hover:bg-gray-50"
           }`}
           key={folder.slug}
-          onClick={() => setActive(folder)}
+          onClick={() => onClick(folder)}
         >
           <LinkIcon />
           {folder.slug}
