+++ src/pages/index.tsx
@@ -1,10 +1,6 @@
-import {
-  PaddleButton,
-  PaddleLoader,
-} from "@/components/integrations/PaddleButton"
 import { Navbar } from "@/components/nav/Navbar"
 import Head from "next/head"
-import { useEffect, useState } from "react"
+import { useEffect } from "react"
 
 export default function Home() {
   useEffect(() => {}, [])
