--- src/utils/router.ts
@@ -1,7 +1,0 @@
-export const router = {
-  home: "/",
-  auth: "/auth",
-  manage: "/manage",
-  pricing: "/pricing",
-  benefits: "/benefits"
-}
